no commit message

+++ Components/BP.WF/HttpHandler/WF_MyCC.cs
@@ -702,7 +702,7 @@ namespace BP.WF.HttpHandler
 
                 #region  加载自定义的button.
                 BP.WF.Template.NodeToolbars bars = new NodeToolbars();
-                bars.Retrieve(NodeToolbarAttr.FK_Node, this.FK_Node,NodeToolbarAttr.ShowWhere, ShowWhere.CC);
+                bars.Retrieve(NodeToolbarAttr.FK_Node, this.FK_Node,NodeToolbarAttr.ShowWhere, (int)ShowWhere.CC);
                 foreach (NodeToolbar bar in bars)
                 {
